Added Test Cases for card functions
These tests don't test for cards being taken out of an 'UnoPlayer's hand, or that the cards were even in the player's hand in the first place. These tests only check for the update of state of the UnoGame, and that the 'Draw Two' Cards and 'Wild Draw Four' Cards do indeed add cards to the next player in the UnoGame. I also made some minor changes in UnoGame.ts to fix any test cases that were failed.